More doc and cleanup
