+++ benchmarks/functional_autograd_benchmark/functional_autograd_benchmark.py
@@ -11,6 +11,7 @@ import audio_text_models
 
 from utils import get_str
 
+# Listing of the different tasks
 FAST_TASKS_NO_DOUBLE_BACK = [
     "vjp",
 ]
@@ -28,6 +29,13 @@ ALL_TASKS = FAST_TASKS + [
 
 DOUBLE_BACKWARD_TASKS = ["jvp", "hvp", "vhp", "hessian"]
 
+# Model definition which contains:
+# - name: a string with the model name.
+# - getter: a function to get the mode. It takes as input the device on which the model
+#     will run. It should return the forward function and the parameters (Tensors) used as
+#     input for the forward function. Note that the forward must *not* have any side effect.
+# - tasks: the list of recommended tasks that can run in a reasonable amount of time with this model.
+# - unsupported: the list of tasks that this model cannot run.
 ModelDef = namedtuple("ModelDef", ["name", "getter", "tasks", "unsupported"])
 
 MODELS = [
@@ -36,7 +44,6 @@ MODELS = [
     ModelDef("detr", vision_models.get_detr, FAST_TASKS, []),
     ModelDef("ppl_simple_reg", ppl_models.get_simple_regression, ALL_TASKS, []),
     ModelDef("ppl_robust_reg", ppl_models.get_robust_regression, ALL_TASKS, []),
-    # ModelDef("ppl_flakyness", ppl_models.get_test_flakyness, ALL_TASKS, []),
     ModelDef("wav2letter", audio_text_models.get_wav2letter, FAST_TASKS, []),
     ModelDef("deepspeech", audio_text_models.get_deepspeech, FAST_TASKS_NO_DOUBLE_BACK, DOUBLE_BACKWARD_TASKS),
     ModelDef("transfo", audio_text_models.get_transformer, FAST_TASKS, []),
@@ -46,6 +53,7 @@ MODELS = [
 def run_once(model, inp, task, extra=None):
     func = getattr(functional, task)
 
+    # Cache the creation of the `v` argument
     if extra is None:
         if task in ["vjp"]:
             out = model(*inp)
@@ -101,11 +109,15 @@ def main():
     parser.add_argument("--num-iters", type=int, default=10)
     parser.add_argument("--gpu", type=int, default=-2, help="GPU to use, -1 for CPU and -2 for auto-detect")
     parser.add_argument("--run-slow-tasks", action="store_true", help="Run even the slow tasks")
-    parser.add_argument("--model-filter", type=str, default="")
-    parser.add_argument("--task-filter", type=str, default="")
+    parser.add_argument("--model-filter", type=str, default="", help="Only run the models in this filter")
+    parser.add_argument("--task-filter", type=str, default="", help="Only run the tasks in this filter")
+    parser.add_argument("--num-threads", type=int, default=10,
+                        help="Number of concurrent threads to use when running on cpu")
     args = parser.parse_args()
 
     results = defaultdict(defaultdict)
+    torch.set_num_threads(args.num_threads)
+    torch.set_num_interop_threads(args.num_threads)
 
     if args.gpu == -2:
         args.gpu = 0 if torch.cuda.is_available() else -1
