+++ benchmarks/functional_autograd_benchmark/ppl_models.py
@@ -90,6 +90,7 @@ def get_robust_regression(device):
     return forward, (nu_value.to(device), sigma_unconstrained_value.to(device), beta_value.to(device))
 
 
+# TODO: Fix this test when we get feedback from Nazanin. Leaving it here for reference.
 def get_test_flakyness():
     # With predefined N, R, F and Y, where N, R and F are just numbers
     N = 10
